personal-gpt-context-engine/pyproject.toml
[project]
name = "personal-gpt-context-engine"
version = "0.1.0"
description = "This is a personal GPT context engine. This engine responsibility is to mange vector data for the personal GPT application."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.121.2",
    "langchain>=1.0.7",
    "langchain-text-splitters>=1.0.0",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.16.0",
    "redis>=7.0.1",
    "sentence-transformers>=5.1.2",
    "tenacity>=9.1.2",
    "typing>=3.10.0.0",
    "uvicorn>=0.38.0",
]
